winapi: typed handle kinds, privatized state, scene enum
Handle counters become per-kind newtypes via a shared handle_newtype
macro (file, find, resource, module, registry, DC, bitmap, brush, pen,
font, hook, console buffer, kernel object), state structs lose their
pub fields in favor of accessors, FakeVa gets a single encode method,
unwind reads get a real error enum, and the d3d9 scene flag becomes a
SceneState enum. Index loops become iterators where semantics allow.

+++ crates/wie-winapi/src/kernel32/misc/mod.rs
@@ -643,7 +643,7 @@ pub(crate) fn handle_rtl_capture_context(
     }
     // Rflags at +0x44
     if let Some(slot) = cbuf.get_mut(0x44..0x4C) {
-        slot.copy_from_slice(&tctx.rflags.to_le_bytes());
+        slot.copy_from_slice(&u64::from(tctx.rflags).to_le_bytes());
     }
     // XMM0..XMM15 at +0x100
     for i in 0_usize..16 {
